Initial commit.
This is the initial commit of the CIU Stats repo. Currently the program
creates a window with data entry fields, data management buttons, and a
treeview to display the entered data.

The program can't add data to the treeview yet, but the "ADD" button
does add the entered data to a list and clear the fields. None of the
other buttons do anything.